backend/my_env/requirements.txt
Django==5.0.3
django-crispy-forms==2.1
djangorestframework==3.15.1